build: Comply with -Wextra compiler flag

+++ src/tools/irm/irm_utils.c
@@ -28,7 +28,7 @@
 
 int matches(const char * cmd, const char * pattern)
 {
-        int len = strlen(cmd);
+        size_t len = strlen(cmd);
 
         if (len > strlen(pattern))
                 return -1;
